Added script for a wild pokemon
When you meet a wild pokemon, the server will force
a battle between you and a client "Wild Pokemon", setting
the team of "Wild Pokemon" to one wild pokemon.

The wild pokemon client will have that script so it
battles randomly. (it will also have a script to close
the battle window when it's finished...)